Create a simpler starter html file, so people can get demo with their data working more easily
There's a lot to take in with DeepScatter, especially in the `index.html` file. Web dev is my day job, & it took me a good while to weed through it all & figure out what I needed for my own data. This file provides a good place to just see the basics.